ZBUG-1948: Fix for ProblematiEmail.eml for getting
the NullPointerException exception:

The root cause of this problem is that in this particular
MIME there was no dummy html tags were present, verified
by creating a empty mail body without having any contents
with attchment that the dummy <html><body></body></html>
are present, so the html was evaluating tonull and we were
getting the NullPointerException exception.

To fix this issue if the html is null or empty then the dummy
html tags are added.